Usage of the `info` or `debug` level logging when `stdout` is a `TTY` results in a ~25% performance hit.
This PR simply adds a warning so developers are aware.

```bash
stdout is a TTY
    iteration_duration.........: avg=71.83ms min=15.43ms med=70.58ms max=135.37ms p(90)=86.29ms p(95)=92.87ms
    iterations.................: 10000  1388.112538/s

stdout is /dev/null
    iteration_duration.........: avg=55.19ms min=23.44ms med=54.81ms max=81.28ms p(90)=62.97ms p(95)=65.49ms
    iterations.................: 10000  1806.935143/s
```