@orangejulius 2020-03-12
Since this is true everwhere, should we consider putting the warning somewhere in our general documentation?

I'd rather we didn't add code to print this warning, especially in one of the (currently) least performance sensitive Pelias services.